bpf_host: declare trace monitor in to_netdev, pass to ipv6 datapath
For https://github.com/cilium/cilium/issues/12561.

To be able to send aggregated traces from 'to_netdev()', the monitor
variable needs to be declared there and passed to the datapath as an
argument. This is the IPv6 path, IPv4 will follow separately as it
moves some logic to a func.

checkpatch.pl trips over srcID, so that was changed to src_id.

Signed-off-by: Timo Beckers <timo@isovalent.com>