bpf: convert bpf_elf_map definitions to BTF map definitions
This is the first in a series of patches that converts most bpf_elf_map
definitions to BTF map definitions. The changes in this patch are limited
to maps that are supported by both cilium/ebpf and iproute2+libbpf.

Further work is necessary to convert PROG_ARRAY and HASH_OF_MAPS map types
given the legacy (iproute2) that needs to be handled. This patch contains
mostly non-snowflakes; maps that were more involved to convert are
addressed in separate commits.

A __section_maps_btf macro was added to emit BTF map definitions into the
well-known .maps section.

Two other macros, __uint and __type were added, which are part of the BTF
map definition spec/layout. The premise of BTF map definitions is that all
struct fields must be pointers to save space in the .maps section, and
(ab)using the BTF type graph to encode the actual values. In the case of
__uint, the length of the array declared by the macro is set to the value
given to the macro. The length of the array is then queried at load time
and used as the 'value' of that struct member.

A few more changes were required to functions that passed an explicit type
of btf_elf_map *, which was changed to void *, as the BPF kernel-side map
API also takes void *.

Signed-off-by: Timo Beckers <timo@isovalent.com>