test/bpf: elf-demo - convert CALLS_MAP in test suite to BTF map definition
This is an early example of what statically initialized PROG_ARRAY maps
look like. This way of using .values is currently only supported by
cilium/ebpf. While working on this, I noticed this program is never loaded
into the kernel, which should probably be changed later as we can start
making more use of PROG_RUN for testing purposes.

Signed-off-by: Timo Beckers <timo@isovalent.com>