bpf: run docker target for checking testdata in CI
If the Makefile ends up building one of the .c files, it also invokes
llvm-objcopy for stripping the binary, which is not installed on the
GHA runners.

Always use the `docker` target for verifying if testdata was correctly
built, since different versions of the clang toolchain generate different
binaries. It's also important to run a clean build so Make doesn't simply
skip building the objects.

Signed-off-by: Timo Beckers <timo@isovalent.com>